@@ -1,115 +1,45 @@
 {
-  "name": "react-material",
-  "version": "2.0.0",
-  "main": "index.js",
-  "license": "MIT",
+  "name": "paper-dashboard-react",
+  "version": "1.0.0",
+  "private": true,
+  "homepage": "https://creativetimofficial.github.io/paper-dashboard-react/#/dashboard",
   "dependencies": {
-    "@material-ui/core": "^1.0.0",
-    "@material-ui/icons": "^1.1.0",
-    "autosuggest-highlight": "^3.1.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "bootstrap": "^4.0.0",
-    "bootstrap-v4-rtl": "^4.0.0-0",
-    "can-use-dom": "^0.1.0",
-    "classnames": "^2.2.5",
-    "copy-webpack-plugin": "^4.2.0",
-    "downshift": "^1.28.0",
-    "draft-js": "^0.10.4",
-    "draftjs-to-html": "^0.8.1",
-    "firebase": "^5.0.2",
-    "history": "^4.7.2",
-    "install": "^0.10.1",
-    "isomorphic-fetch": "^2.2.1",
-    "jss-rtl": "^0.2.3",
-    "material-ui-pickers": "^1.0.0-rc.11",
-    "moment": "^2.19.2",
-    "node-sass": "^4.7.2",
-    "nprogress": "^0.2.0",
-    "package.json": "^2.0.1",
-    "path": "^0.12.7",
-    "prop-types": "^15.6.0",
-    "react": "^16.4.1",
-    "react-autosuggest": "^9.3.3",
-    "react-big-calendar": "^0.17.0",
-    "react-bootstrap-sweetalert": "^4.1.0",
-    "react-ckeditor-component": "^1.0.7",
-    "react-color": "^2.13.8",
-    "react-custom-scrollbars": "^4.2.1",
-    "react-device-detect": "^0.1.0",
-    "react-dnd": "^2.5.4",
-    "react-dnd-html5-backend": "^2.5.4",
-    "react-dom": "^16.4.1",
-    "react-draft-wysiwyg": "^1.12.2",
-    "react-dropzone": "^4.2.9",
-    "react-google-maps": "^9.4.5",
-    "react-hot-loader": "^3.1.3",
-    "react-icons": "^2.2.7",
-    "react-intl": "^2.4.0",
-    "react-joyride": "^1.11.0",
-    "react-loadable": "^5.3.1",
-    "react-notifications": "^1.4.3",
-    "react-number-format": "^3.0.2",
-    "react-owl-carousel": "^2.3.0",
-    "react-placeholder": "^2.0.0",
-    "react-redux": "^5.0.6",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-router-redux": "5.0.0-alpha.6",
-    "react-select": "^1.2.1",
-    "react-sortable-hoc": "^0.6.8",
-    "react-star-rating-component": "^1.3.0",
-    "react-swipeable-views": "^0.12.8",
-    "react-text-mask": "^5.0.2",
-    "reactstrap": "^6.2.0",
-    "recharts": "^1.0.0-alpha.7",
-    "recompose": "^0.26.0",
-    "redux": "^3.7.2",
-    "redux-logger": "^3.0.6",
-    "redux-saga": "^0.16.0",
-    "save": "^2.3.1",
-    "script-ext-html-webpack-plugin": "^1.8.8",
-    "style-ext-html-webpack-plugin": "^3.4.7",
-    "typeface-roboto": "^0.0.45"
+    "@types/react": "16.4.16",
+    "bootstrap": "4.1.3",
+    "chart.js": "2.7.2",
+    "history": "4.7.2",
+    "node-sass": "4.9.3",
+    "perfect-scrollbar": "1.4.0",
+    "react": "16.5.2",
+    "react-chartjs-2": "2.7.4",
+    "react-dom": "16.5.2",
+    "react-google-maps": "9.4.5",
+    "react-notification-alert": "0.0.8",
+    "react-router-dom": "4.3.1",
+    "react-scripts": "2.0.4",
+    "reactstrap": "6.5.0"
   },
   "scripts": {
-    "clean": "rimraf dist/*",
-    "build": "webpack --progress --bail --env dist -p",
-    "start": "webpack-dev-server --hot --open"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject",
+    "scss": "node-sass --include-path ./src --include-path ./node_modules src/ -o src/",
+    
+    "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && npm start"
   },
-  "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-istanbul": "^4.1.4",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-object-rest-spread": "^6.23.0",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "css-loader": "^0.28.5",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
-    "eslint": "^4.6.0",
-    "eslint-config-airbnb": "^15.0.1",
-    "eslint-loader": "^1.9.0",
-    "eslint-plugin-flowtype": "2.34.0",
-    "eslint-plugin-import": "^2.3.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.0.1",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^0.11.2",
-    "html-webpack-plugin": "^2.30.1",
-    "less": "^2.7.3",
-    "less-loader": "^4.0.5",
-    "postcss-loader": "^2.0.8",
-    "sass-lint": "^1.11.1",
-    "sass-loader": "^6.0.6",
-    "style-loader": "^0.18.2",
-    "webpack": "^3.5.6",
-    "webpack-dev-server": "^2.7.1"
-  },
-  "engines": {
-    "node": ">=6.9.0",
-    "npm": ">= 3"
+ 
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ],
+  "optionalDependencies": {
+    
+    "ajv": "6.0.0",
+    "@types/googlemaps": "3.30.13",
+    "@types/markerclustererplus": "2.1.33",
+    "jquery": "3.3.1"
   }
 }
